Add helper to find current client row
Many of the actions, like get/task etc require that the client is registered,
and "active" in the db.

So create a common helper that uses the ua, ip and guid/sessionid params to
find that DBIx::Class client object, else return undef, letting the controller
decided what kinda error to throw.