On api error, stop trying to work
When we get an error condition back from the API, we call update_state() to
ensure the buttons are reflective of the current server state.  This will also
stop us doing any more work if the server says we're no longer connected.

And thus, only poll for more work if we got a non-error.