Show details about a specific task & add 'array' helper
http://192.168.58.100:8899/client/task/149

* created, started, finished times
* which clients/browsers ran the test, and what the results were

re: 'array' helper.

Using Xslate, DBIx::Class relationships are always treated in scalar context by
the template, meaning DBIx::Class always returns the count, not the list.

    [% FOR ct IN task.client_tasks %]

will always return a scalar int for task.client_tasks, not the array ref we
want, regardless of how you call it in the template.  Aka, it's equivalent to:

    [% FOR ct IN 4 %]

WHich isnt' what we want.

Thus, we introduce the 'helper' called 'array' which basically forces things
into an array for us.

    [% FOR ct IN array( task.client_tasks ) %]

It's icky, but it works, and tho it's a few extra boilerplate chars in the
template, it does help make the intent explicit.