Correctly use .epoch and avoid undef error on task/list
1. started_at and finished_at are DateTime (auto inflated) objects, that don't
support simple - operations.  What we *meant* was to use epoch, which the raw
underlying db column already is, but by the time we're in the template, it's
now been converted into a DateTime object.  So convert it back ;)

2. Don't try and do finished_at.epoch - started_at.epoch unless both are
actually defined.  It's at least in theory possible to get a started_at but no
finished_at.