Make the app's base configurable.
---

Can't find a better way to get the parent (lib/Vulture.pm) to stash a value the
templates can see.  $self->stash() seems to get cleared out.

Using a function works, but it's ugly.

---

In theory, the only page that *need* to be served by the app, and done when in
"proxy" mode, is the one that first loads the initial client test page, that
will contain the iframe with the target test url/page.  Required so as to get
around the XSS restrictions.

We could then choose to hardwire all api calls to, say,
http://testing.example.com

And ensure the proxy is clever enough to (a) intercept that domain, and (b)
re-direct the request back to itself.

However I think we'd still want that intercept domain to be configurable, and
for now, simply making all local request paths use some configurably rare
prefix ought to suffice.

What you don't want is:

    $api_base/some/local/api/path

matching a legit request that we instead should be proxypassing.