Add "count" and "test_name" cols to job_result table
We get something like:

    ok 1 name of test
    not ok 2 name of other test
    # some diag text

For the none diag lines, we want to split out:

    - state (ok/not ok)
    - count/number
    - test name

Which for now, we achieve with some ugly regex parsing of the raw tap result
line.  In due course, it would be nicer if the client tracked these fields
seperately, and submitted them separately.

Primarily, this would give us the ability to associate specific "diag" lines
with a specific "test_name".  For example, rather than:

    top.like(got, /some regex/, 'test for some regex');

when failing producing:

    not ok 1 'test for some regex'
    # Got 'another regex'
    # Expected /some regex/

as three unconnected lines, we'd instead like something akin to:

    var results = [
        {
            'state'     : 'not ok',
            'count'     : 1,
            'test_name' : 'test for some regex',
            'diag'      : [
                "Got 'another regex'",
                "Expected /some regex/",
            ]
        }

Which we would pass back to the server to log/store.  We can still take that
and convert it back to the native/naive TAP multiline output, but now we can
also know that the diag lines are associated with that specific test.

Thus, when later reviewing test results, for failing tests we can also show the
associated diag lines.