Auto redirect task/edit to correct domain for current task
When we first load, ensure the browser domain matches the test url domain and
if not, redirect.  We will have just loaded, urgo it's safe to save the test.

XXX Do we need to save when first loading?

When we get a chance on the test url input field, check if we need to redirect.

If we do, confirm it's okay to save the test (required to ensure the new test
url is saved across the redirect, and to avoid loosing changes to the test js)

Then redirect, so that the current browser domain matches our new test url.