Return job, not error, if api/run/get finds 'running' tasks.
Consider:

    - Load HQ on http://example.com/...
    - Starts working ....
    - Sent test for http://foo.com/

At this point we want HQ to reload itself on the new domain, preserving the
existing app/client ids.

Having reloaded, we'll request work to do.

The current server side state tells us this client has already been given a
job, which was the one that initiated the HQ page reload, and which won't have
been run yet.

So rather than returning an error saying "that client already has a task",
instead we return the oldest job for this client that's marked as being in
state 'running'.  Which hopefully, is the same job we originally gave it, that
caused the reload.

In short, this should ensure the client resumes the existing job after a
page-reload, which we've done to facilitate the change of domain.