Auto redirect HQ to correct domain for current job/task
We're initially loaded on

    example.com

then we get an incoming test for:

    other.com

we need to have the page reload itself onto the correct domain for the test
(other.com).

To do this, we pass through the client & app id, which ensures after we've
reloaded ourselves on the new domain that these values are retained and not
reset.  Note that the app_id is stored on localStorage, but this is keyed on
the current domain, hence why we still need to pass this through.

Finally, we tell ourselves to restart working.  Because we haven't explicitly
disconnected, the server will still consider our current client to be active.

Hence we simply need to ask the server for more work.  The server should be
smart enough to re-send the previous (running) job, that was the original cause
of having to reload the page onto the new/correct domain.